requirements.txt
boto3==1.17.70
botocore==1.20.70
install==1.3.4
jmespath==0.10.0
numpy==1.20.3
opencv-python==4.5.2.52
pdf2image==1.14.0
Pillow==8.2.0
pytesseract==0.3.7
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.4
